+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3890,7 +3890,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01M36FZE9S44D3B53RGRCQDJ7T"),
+  id: z.string().default("01M36HS44XZ9GGYYNC805AB4T8"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),